Move relevant files to Durable\Commands folder
