+++ src/Durable/Commands/StopDurableTimerTaskCommand.cs
@@ -5,7 +5,7 @@
 
 #pragma warning disable 1591 // Missing XML comment for publicly visible type or member 'member'
 
-namespace Microsoft.Azure.Functions.PowerShellWorker.Durable
+namespace Microsoft.Azure.Functions.PowerShellWorker.Durable.Commands
 {
     using System.Management.Automation;
